This commit primarily addresses the way data is organized as Causal Mask Modeling. The approach we previously used, Flan Encoder Decoder, may have been incorrect. Shout out to @ZhangYuanhan-AI 

Former-commit-id: 212619ec0d30e46e97f4899e5c1f9b2d7412577e

+++ open_flamingo/src/flamingo.py
@@ -91,10 +91,11 @@ class Flamingo(nn.Module):
             # Case: do not use caching (i.e. this is a standard forward pass);
             self._encode_vision_x(vision_x=vision_x)
 
+        # import pdb;pdb.set_trace()
         output = self.lang_encoder(
-            input_ids=lang_x,
+            input_ids=lang_x.int(),
             attention_mask=attention_mask,
-            labels=labels,
+            labels=labels.long(),
             past_key_values=past_key_values,
             use_cache=use_cache,
         )
